Configurable channel name template
Allow a playbook to define a template for the playbook run channel name,
pre-populating the interactive dialog with this value when the playbook
is run from the RHS List or the backstage.

Due to limitations of interactive dialogs at present, this experience
doesn't work with `/playbook start`, as there are multiple playbook
options and no way to update the run name after selecting one.
Fortunately, this path is not exposed in the UI outside of slash
commands.